Update readme to reflect current state of Cloudera-Deploy at public cutover.
The user profile is the first `include` statement executes, whether that means additional `include` statements will not overwrite that key is not currently tested.
The `plat` tag does not yet include CDP Private Base deployments, they are still under `full_cluster` and `default_cluster` at this time.
A user should not set `globals` directly, as it is composed during the init role process from supplied top level keys in the profile and elsewhere.

Signed-off-by: Daniel Chaffelson <chaffelson@gmail.com>